Package reorganisation
- Moved atmosphere child modules
- Removed unused code, cleaned up some imports
- Renamed lithosphere to surface
