+++ docs/rst/api_reference/radprops.rst
@@ -1,0 +1,20 @@
+.. _sec-api_reference-radprops:
+
+Radiative properties [eradiate.radprops]
+========================================
+
+.. rubric:: Description
+
+.. automodule:: eradiate.radprops
+
+.. currentmodule:: eradiate.radprops
+
+.. rubric:: Modules
+
+.. autosummary::
+   :recursive:
+   :toctree: generated/
+
+   absorption
+   rad_profile
+   rayleigh
