+++ docs/rst/api_reference/thermoprops.rst
@@ -1,0 +1,19 @@
+.. _sec-api_reference-thermophysics:
+
+Thermophysical properties [eradiate.thermoprops]
+================================================
+
+.. rubric:: Description
+
+.. automodule:: eradiate.thermoprops
+
+.. currentmodule:: eradiate.thermoprops
+
+.. rubric:: Modules
+
+.. autosummary::
+   :recursive:
+   :toctree: generated/
+
+   us76
+   util
