+++ eradiate/thermoprops/__init__.py
@@ -80,5 +80,3 @@
 
     (coming).
 """
-
-from . import us76
