+++ eradiate/util/collections.py
@@ -1,8 +1,6 @@
 """Specialised container datatypes providing alternatives to Python’s general
 purpose built-in containers, dict, list, set, and tuple."""
 
-import collections
-
 from dpath import util as dpu
 from dpath.exceptions import PathNotFound
 
@@ -14,69 +12,6 @@ def onedict_value(d):
         raise ValueError(f"dictionary has wrong length (expected 1, got {len(d)}")
 
     return next(iter(d.values()))
-
-
-class frozendict(collections.abc.Mapping):
-    """A frozen dictionary implementation. See
-    https://stackoverflow.com/questions/2703599/what-would-a-frozen-dict-be.
-
-    It behaves like a dictionary, except that it cannot be modified after
-    initialisation.
-    """
-
-    def __init__(self, *args, **kwargs):
-        self._d = dict(*args, **kwargs)
-        self._hash = None
-
-    def __iter__(self):
-        return iter(self._d)
-
-    def __len__(self):
-        return len(self._d)
-
-    def __getitem__(self, key):
-        return self._d[key]
-
-    def __repr__(self):
-        return f"frozendict({self._d.__repr__()})"
-
-    def __hash__(self):
-        # It would have been simpler and maybe more obvious to
-        # use hash(tuple(sorted(self._d.iteritems()))) from this discussion
-        # so far, but this solution is O(n). I don't know what kind of
-        # n we are going to run into, but sometimes it's hard to resist the
-        # urge to optimize when it will gain improved algorithmic performance.
-        if self._hash is None:
-            hash_ = 0
-            for pair in self.items():
-                hash_ ^= hash(pair)
-            self._hash = hash_
-        return self._hash
-
-    def copy(self):
-        """Return shallow copy of encapsulated dict"""
-        return self._d.copy()
-
-
-def update(d, u):
-    """This function updates nested dictionaries.
-    See https://stackoverflow.com/questions/3232943/update-value-of-a-nested-dictionary-of-varying-depth.
-
-    Parameter ``d`` (dict)
-        Dictionary which will be updated (**modified in-place**).
-
-    Parameter ``u`` (dict)
-        Dictionary holding the values to update ``d`` with.
-
-    Returns → dict
-        Updated ``d``.
-    """
-    for k, v in u.items():
-        if isinstance(v, collections.abc.Mapping):
-            d[k] = update(d.get(k, {}), v)
-        else:
-            d[k] = v
-    return d
 
 
 class ndict(dict):
